bug 1042820 (part) migrate mail sending functionality from apps/amo to mkt/site
