--- apps/amo/templates/amo/emails/unsubscribe.html
@@ -1,10 +1,0 @@
-{{ message }}
-{% if show_unsubscribe %}
-
-<p>
-===================================================================<br>
-You received this email because you're subscribed to:
-</p>
-<ul>
-   <li>email me when {{ perm_setting }}</li>
-</ul>
