+++ mkt/webapps/models.py
@@ -35,7 +35,7 @@ from amo.decorators import skip_cache, use_master, write
 from amo.helpers import absolutify
 from amo.storage_utils import copy_stored_file
 from amo.utils import (attach_trans_dict, find_language, JSONEncoder,
-                       send_mail, slugify, smart_path, sorted_groupby, timer,
+                       slugify, smart_path, sorted_groupby, timer,
                        to_language, urlparams)
 from lib.crypto import packaged
 from lib.iarc.client import get_iarc_client
@@ -50,6 +50,7 @@ from mkt.files.utils import parse_addon, WebAppParser
 from mkt.prices.models import AddonPremium, Price
 from mkt.ratings.models import Review
 from mkt.regions.utils import parse_region
+from mkt.site.mail import send_mail
 from mkt.site.models import DynamicBoolFieldsMixin
 from mkt.tags.models import Tag
 from mkt.translations.fields import (PurifiedField, save_signal,
