+++ mkt/webpay/views.py
@@ -17,7 +17,6 @@ from rest_framework.viewsets import GenericViewSet
 
 import amo
 from amo.helpers import absolutify, urlparams
-from amo.utils import send_mail_jinja
 from lib.cef_loggers import app_pay_cef
 from mkt.api.authentication import (RestAnonymousAuthentication,
                                     RestOAuthAuthentication,
@@ -26,6 +25,7 @@ from mkt.api.authorization import AllowReadOnly, AnyOf, GroupPermission
 from mkt.api.base import CORSMixin, MarketplaceView
 from mkt.purchase.models import Contribution
 from mkt.receipts.utils import create_inapp_receipt
+from mkt.site.mail import send_mail_jinja
 from mkt.webpay.forms import FailureForm, PrepareInAppForm, PrepareWebAppForm
 from mkt.webpay.models import ProductIcon
 from mkt.webpay.serializers import ProductIconSerializer
