+++ mkt/zadmin/tasks.py
@@ -4,7 +4,7 @@ from django.conf import settings
 
 from celeryutils import task
 
-from amo.utils import send_mail
+from mkt.site.mail import send_mail
 from mkt.zadmin.models import EmailPreviewTopic
